+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add: #27 RegEx filter `regex` for excerpt
 - Add: Option for logfile
 - Add: #28 Stripe note-overview code block fields
+- Add: #26 Date variables for search query
 
 ## v1.5.5 (2022-01-26)